requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-chroma
sentence-transformers
huggingface-hub
tqdm
groq
